+++ hooks/pages/index.ts
@@ -3,3 +3,5 @@ export { useRecipientDetailPage } from './useRecipientDetailPage';
 export { useRecipientsListPage } from './useRecipientsListPage';
 
 export { useNewRecipientForm } from './useNewRecipientForm';
+
+export { useMessageForm } from './useMessageForm';
